+++ Makefile
@@ -1,26 +1,25 @@
 # Define Go command and flags
 GO = go
 GOFLAGS = -ldflags="-s -w"
-
-# Define the target executable
 TARGET = gogoodwe
+MAINAPPPATH = ./cmd/${TARGET}/main.go
 
 ## help - Display help about make targets for this Makefile
 help:
 	@cat Makefile | grep '^## ' --color=never | cut -c4- | sed -e "`printf 's/ - /\t- /;'`" | column -s "`printf '\t'`" -t
 
 ## localrelease -  Builds the project in preparation for (local)release
-localrelease: vet lint staticcheck seccheck
-	go build $(GOFLAGS) -o bin/${TARGET} main.go
+localrelease: vet lint seccheck unittest
+	go build $(GOFLAGS) -o bin/${TARGET} ${MAINAPPPATH}
 	file bin/${TARGET}
 
 ## release - Builds the project in preparation for release
 release:
-	goreleaser release --snapshot --clean
+	goreleaser release --snapshot --clean 
 	
 ## debug - Builds the project in preparation for debug
 build:
-	go build -o bin/${TARGET} main.go
+	go build -o bin/${TARGET} ${MAINAPPPATH}
 	file bin/${TARGET}
 
 ## buildandrun - builds and runs the program on the target platform
@@ -29,7 +28,7 @@ buildandrun: build
 
 ## run - runs main.go for testing
 run: dep
-	go run main.go
+	go run ${MAINAPPPATH}
 
 
 ## clean - Remove the old builds and any debug information
@@ -38,9 +37,9 @@ clean:
 	rm -rf dist
 	rm bin/${TARGET}
 
-## test - executes unit test
-test:
-	go test ./...
+## unittest - executes unit tests
+unittest:
+	go test -v ./test/...
 
 ## dep - fetches any external dependencies
 dep:
@@ -51,9 +50,9 @@ dep:
 vet:
 	go vet ./...
 
-## staticcheck - Runs static code analyzer staticcheck
-staticcheck: 
-	go run honnef.co/go/tools/cmd/staticcheck@latest -checks=all,-ST1000,-U1000 ./...
+## staticcheck - Runs static code analyzer staticcheck - currently broken
+# staticcheck: 	
+# 	go run honnef.co/go/tools/cmd/staticcheck@latest ./...
 
 ## seccheck - Code vulnerability check
 seccheck:
