+++ test/samplemodule_test.go
@@ -1,0 +1,22 @@
+package testing
+
+/*
+A Sample test harness.
+*/
+
+// import (
+// 	"testing"
+
+// 	"github.com/AaronSaikovski/gostarter/internal/pkg/samplemodule"
+// )
+
+// // A testing function.
+// func TestSampleFunction(t *testing.T) {
+
+// 	msg := samplemodule.SampleFunction()
+// 	expected := "OK"
+
+// 	if msg != expected {
+// 		t.Errorf("Module expected '%q' but got '%q'", expected, msg)
+// 	}
+// }
