+++ test/samplestruct_test.go
@@ -1,0 +1,21 @@
+/*
+A Sample test harness.
+*/
+
+package testing
+
+// import (
+// 	"github.com/AaronSaikovski/gostarter/internal/app/types"
+// 	"testing"
+// )
+
+// // A testing function.
+// func TestSampleStructString(t *testing.T) {
+
+// 	expected := "test data"
+// 	ateststruct := types.Sample{SampleString: "test data", SampleInt: 1}
+
+// 	if ateststruct.SampleString != expected {
+// 		t.Errorf("struct expected '%s' but got '%s'", expected, ateststruct.SampleString)
+// 	}
+// }
